+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.1.5
+  - The server can now do client side verification by providing a list of certificate authorities and configuring the `ssl_verify_mode`,
+    the server can use `peer`, if the client send a certificate it will be validated. Using `force_peer` will make sure the client provide a certificate
+    and it will be validated with the know CA.  #8
 # 2.1.4
   - Change the `logger#warn` for `logger.debug` when a peer get disconnected, keep alive check from proxy can generate a lot of logs  #46
 # 2.1.3